requirements.txt
git+https://github.com/KurimuzonAkuma/pyrogram.git@v2.1.
tgcrypto
python-dotenv
pymongo
